default gateway support
